Add support for outputting the non zeros in L and U.  Add the support of LongInt
The output is given in PMatrix which is independent of the factorization routine used (partly because it is relatively difficult to hack into other factorization routines).

The LongInt is supported in the level of environment.hpp,

which uses int64_t in stdint.h and this is platform independent.

The total number of non zeros in PMatrix is given by the LongInt format.