Change new_ppexsi and new_interface to ppexsi and interface.
The templateinterf branch is now essentially finished and is ready to be
merged with the master branch.

There are several routines in the interface that have not been
implemented yet, such as in interface and examples.  These will be
filled in in the future.

Still keep the *old files for future development purpose.