Fix a bug for deallocating SuperLUGrid.
In the PPEXSI driver, the SuperLUGrid is first deallocated explicitly,
and then the luRealMat and luComplexMat are deallocated implicitly.
This is problematic, since the process of deallocating SuperLUMatrix
still need the existence of the SuperLUGrid.

In the fixed version, the luRealMat_, luComplexMat_ are defined as
pointers, and are deallocated explicitly **before** the deallocation of
SuperLUGrid.  For consistency, PMRealMat and PMComplexMat are also
defined as pointers, and are constructed and destructed explicitly.