Patch suggested by Alberto Garcia.
Fix a bug in the "lateral expansion" code for the bracket handling
for inertia-counting in ppexsi.cpp.

Consider an example (such as the h2o molecule) for which the inertia
count at one of the extremes of the interval (say, muMaxInertia) is
exactly 8. Then there should be an expansion on that side. And in fact
the code does that, but at the same time sets muMinInertia to
muMaxInertia, thus triggering again the same problem at the other
side.