Add pole updating strategy.
o DFTDriver2 is added for using the same Green's functions but with
different pole expansions for multiple times.

o driver_ksdft_2.c demonstrates the applicability of the method. This
has been tested both for the metallic and the insulating system. For the
metallic system, it is important to choose muInertiaTolerance to be
on the order of the temperature.

o In DFTDriver2, CalculateFermiOperatorReal2 routine is used. Strictly
one PEXSI calculation per SCF is done. No more Newton iteration.

o The C interface is added. FORTRAN interface and new parameters for the
new strategy TBD.